Move server only functions to actions.ts
